refactor: remove the "god" hook
