+++ apps/web/src/components/event-calendar/hooks/index.ts
@@ -5,15 +5,12 @@
  * Organized by functionality for better maintainability.
  */
 
-// Core state management hooks
-export { useCalendarStateManagement } from "./use-calendar-state-management";
-
-// Individual focused hooks
+// Focused logic hooks
 export { useCalendarNavigation } from "./use-calendar-navigation";
 export { useEventDialog } from "./use-event-dialog";
 export { useEventOperations } from "./use-event-operations";
 export { useKeyboardShortcuts } from "./use-keyboard-shortcuts";
 
-// Utility hooks (existing)
+// Utility hooks
 export { useCurrentTimeIndicator } from "./use-current-time-indicator";
 export { useEventVisibility } from "./use-event-visibility";
